Separated api controller from html controller
