+++ server.js
@@ -1,6 +1,7 @@
 var express = require("express");
 var bodyParser = require("body-parser");
 var router = require("./controllers/html-controller");
+var router2 = require("./controllers/events_controller")
 var session = require("express-session");
 var passport = require("./config/passport");
 
@@ -29,6 +30,8 @@ app.use(passport.session());
 require("./controllers/user-controller.js")(app);
 
 app.use(router);
+
+app.use(router2);
 
 
 // Start our server so that it can begin listening to client requests.
